requirements.txt
antiorm==1.2.1
asgiref==3.7.2
DateTime==5.2
db==0.1.1
dj-database-url==2.1.0
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.0
PyJWT==2.8.0
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
zope.interface==6.0
